qoqo_for_braket/pyproject.toml
[project]
name = "qoqo_for_braket"
version = "0.2.6"
license = {file="LICENSE"}
authors = [
    {name="HQS Quantum Simulation GmbH", email="info@quantumsimulations.de"}
]
maintainers = [
    {name="Kirsten Bark", email="kirsten.bark@quantumsimulations.de"}
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "qoqo>=1.4",
    "qoqo_qasm>=0.8.3",
    "amazon-braket-sdk>=1.5.9"
]

[project.optional-dependencies]
tests = [
  'pytest-cov',
  'pytest'
]
dev = [
    'ruff',
    'darglint',
    'bandit',
    'mypy',
    'black'
]
docs = [
    "sphinx >= 2.1",
    "nbsphinx",
    "pygments",
    "recommonmark",
    "myst_parser",
    "sphinx_rtd_theme",
    "tomli"
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 99  # same as black
target-version = "py39"
show-fixes = true
show-source = true
# activate the following checks
select = [
    "A",  # builtins
    "ANN",  # annotations
    "ARG",  # unused arguments
    "B",  # bugbear
    "C",  # comprehensions
    "C90",  # mccabe complexity
    "D",  # pydocstyle
    "E",  # pycodestyle
    "ERA",  # remove commented out code
    "F",  # pyflakes
    "NPY",  # numpy
    "PL",  # pylint
    "RUF",  # ruff
    "S",  # bandit
    "TCH",  # type checking
    "W",  # Warnings
]
# ignore specific violations
ignore = [
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN102",
    "ANN401",
    "D400",
    "D401",
    # "S403",  # not implemented in ruff (yet?!)
    # "S404",  # not implemented in ruff (yet?!)
    "PLR",
    "S301",
]
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W",
    "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT",
    "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI",
    "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"
]
unfixable = []
# exclude the following patterns from linting
exclude = [
    ".git",
    "__pycache__",
    "doc/conf.py",
    "old",
    "build",
    "dist",
    "test_*"
]

[tool.ruff.mccabe]
max-complexity = 20  # 5 higher than sonarqube

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]


[tool.mypy]
ignore_missing_imports = true
pretty = true
follow_imports = "silent"

[tool.bandit]

[tool.black]
line-length = 99
target-version = ['py39', 'py310', 'py311']
